+++ apps/h5/src/enhance.ts
@@ -1,0 +1,6 @@
+import type { App } from 'vue';
+import type { Provider } from '@vtj/renderer';
+
+export default function (app: App, provider: Provider) {
+  console.log('VtjEnhance app', app, provider);
+}