+++ packages/designer/src/framework/engine.ts
@@ -39,7 +39,8 @@ import {
   type PageFile,
   type HistoryItem,
   type HistoryModelEvent,
-  type BlockSchema
+  type BlockSchema,
+  type EnhanceConfig
 } from '@vtj/core';
 import {
   type Context,
@@ -48,7 +49,7 @@ import {
   Access,
   type ProvideAdapter
 } from '@vtj/renderer';
-import { logger } from '@vtj/utils';
+import { logger, isBoolean } from '@vtj/utils';
 
 import { SkeletonWrapper, type SkeletonWrapperInstance } from '../wrappers';
 import { depsManager, widgetManager } from '../managers';
@@ -132,6 +133,11 @@ export interface EngineOptions {
    * 适配远程接口
    */
   openApi?: OpenApi;
+
+  /**
+   * 开启应用整强
+   */
+  enhance?: boolean | EnhanceConfig;
 }
 
 export const SAVE_BLOCK_FILE_FINISH = 'SAVE_BLOCK_FILE_FINISH';
@@ -176,7 +182,8 @@ export class Engine extends Base {
       access,
       remote = REMOTE,
       checkVersion = true,
-      openApi
+      openApi,
+      enhance = false
     } = this.options;
     this.container = container;
     this.service = service;
@@ -197,7 +204,8 @@ export class Engine extends Base {
     this.assets = new Assets(this.service, this.provider);
     this.simulator = new Simulator({
       engine: this,
-      materialPath
+      materialPath,
+      enhance: this.getEnhanceConfig(enhance)
     });
     this.access = access || new Access({ alert, ...ACCESS });
     this.remote = remote;
@@ -224,6 +232,31 @@ export class Engine extends Base {
       this.saveMaterials();
       this.triggerReady();
       this.report.init();
+    }
+  }
+  private getEnhanceConfig(enhance: boolean | EnhanceConfig) {
+    if (enhance) {
+      const pathname = location.pathname;
+      return isBoolean(enhance)
+        ? {
+            name: 'VTJEnhance',
+            urls: [
+              `${pathname}enhance/enhance.css`,
+              `${pathname}enhance/enhance.umd.js`
+            ]
+          }
+        : enhance;
+    }
+  }
+  private checkLocked(slient?: boolean) {
+    const lockedBy = this.project.value?.locked;
+    const isLocked = !!lockedBy;
+    const info = this.access?.getData();
+    if (isLocked && info?.name !== lockedBy) {
+      if (!slient) {
+        alert(`项目已被[ ${lockedBy} ]锁定，无法更新`);
+      }
+      return true;
     }
   }
 
@@ -282,6 +315,7 @@ export class Engine extends Base {
   }
 
   private async changeFile(e: BlockModel) {
+    if (this.checkLocked()) return;
     await nextTick();
     const dsl = e.toDsl();
     this.service.saveFile(dsl, this.project.value?.toDsl());
@@ -313,6 +347,7 @@ export class Engine extends Base {
   }
 
   private async saveProject(e: ProjectModelEvent) {
+    if (this.checkLocked()) return;
     const project = e.model;
     const dsl = project.toDsl();
     await this.service.saveProject(dsl, e.type);
@@ -320,6 +355,7 @@ export class Engine extends Base {
   }
 
   private async saveBlockFile(e: ProjectModelEvent) {
+    if (this.checkLocked()) return;
     const type = e.type;
     const project = e.model;
     const projectDsl = project.toDsl();
@@ -393,6 +429,7 @@ export class Engine extends Base {
   }
 
   private saveCurrentFile() {
+    if (this.checkLocked()) return;
     const current = this.current.value;
     if (current) {
       this.updateCurrent(current);
@@ -414,6 +451,7 @@ export class Engine extends Base {
   }
 
   private async saveHistory(e: HistoryModelEvent) {
+    if (this.checkLocked(true)) return;
     const type = e.type;
     const history = e.model;
     const projectDsl = this.project.value?.toDsl();
@@ -463,6 +501,7 @@ export class Engine extends Base {
   }
 
   private async publish() {
+    if (this.checkLocked()) return;
     const project = this.project.value;
     if (project) {
       const dsl = {
@@ -477,6 +516,7 @@ export class Engine extends Base {
   }
 
   public async genSource() {
+    if (this.checkLocked()) return;
     const project = this.project.value;
     if (project) {
       const dsl = {
@@ -493,6 +533,7 @@ export class Engine extends Base {
   }
 
   private async publishCurrent() {
+    if (this.checkLocked()) return;
     const project = this.project.value;
     const current = project?.currentFile;
     if (project && current) {