+++ packages/parser/src/vue/index.ts
@@ -50,7 +50,8 @@ export async function parseVue(options: IParseVueOptions) {
   const { nodes, slots, context } = parseTemplate(id, name, sfc.template, {
     platform,
     handlers,
-    styles
+    styles,
+    imports
   });
 
   const dsl: BlockSchema = {
@@ -223,12 +224,19 @@ function parseDeps(
     {} as Record<string, string>
   );
   for (const { from, imports: names } of imports) {
-    names.forEach((name) => {
+    if (Array.isArray(names)) {
+      names.forEach((name) => {
+        const library = depsMap[from];
+        if (library) {
+          libs[name] = library;
+        }
+      });
+    } else {
       const library = depsMap[from];
       if (library) {
-        libs[name] = library;
+        libs[names] = library;
       }
-    });
+    }
   }
   return {
     libs