+++ packages/renderer/src/provider/index.ts
@@ -1,3 +1,4 @@
 export * from './provider';
 export * from './apis';
 export * from './defaults';
+export * from './enhance';