+++ platforms/pro/ide/views/preview.vue
@@ -11,14 +11,19 @@
     Extension,
     createAdapter,
     createServiceRequest,
-    setupPageSetting
+    setupPageSetting,
+    loadEnhance,
+    type VTJConfig,
+    type EnhanceConfig
   } from '../../src';
   import { IconsPlugin } from '@vtj/icons';
   import { useTitle } from '@vueuse/core';
   import { notify, loading, alert } from '../utils';
 
   const service = new LocalService(createServiceRequest(notify));
-  const config = await service.getExtension().catch(() => null);
+  const config: VTJConfig =
+    (await service.getExtension().catch(() => null)) || {};
+  const enhance = await loadEnhance(config.enhance as EnhanceConfig);
   const adapter = createAdapter({
     loading,
     notify,
@@ -28,9 +33,11 @@
   });
   const options = config ? await new Extension(config).load() : {};
   const { __BASE_PATH__ = '/' } = config || {};
+
   const { provider, onReady } = createProvider({
     mode: ContextMode.Runtime,
     service,
+    enhance,
     materialPath: __BASE_PATH__,
     adapter: {
       ...adapter,