+++ apps/app/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.20](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.12.19...vtj-project-app@0.12.20) (2025-05-15)
+
+
+### Features
+
+* ✨ 支持应用增强配置 ([7abb434](https://gitee.com/newgateway/vtj/commits/7abb4349377ef268d6ff7e92691384c0d7980214))
+
+
+
+
+
+## [0.12.19](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.12.18...vtj-project-app@0.12.19) (2025-05-12)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
 ## [0.12.18](https://gitee.com/newgateway/vtj/compare/vtj-project-app@0.12.17...vtj-project-app@0.12.18) (2025-05-12)
 
 **Note:** Version bump only for package vtj-project-app