+++ packages/uni/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.20](https://gitee.com/newgateway/vtj/compare/@vtj/uni@0.12.19...@vtj/uni@0.12.20) (2025-05-15)
+
+**Note:** Version bump only for package @vtj/uni
+
+
+
+
+
+## [0.12.19](https://gitee.com/newgateway/vtj/compare/@vtj/uni@0.12.18...@vtj/uni@0.12.19) (2025-05-12)
+
+**Note:** Version bump only for package @vtj/uni
+
+
+
+
+
 ## [0.12.18](https://gitee.com/newgateway/vtj/compare/@vtj/uni@0.12.17...@vtj/uni@0.12.18) (2025-05-12)
 
 **Note:** Version bump only for package @vtj/uni